\name{DataFreq}

\alias{DataFreq}

\docType{data}

\title{Frequency data set.}

\description{Simulated data set with the weekly frequency of the number of coffee cups consumed weekly in some world capitals.}

\usage{data(DataFreq)}

\format{Set of data with 6 rows and 9 columns. There are 6 observations described by 9 variables: Group by sex and age, Sao Paulo - Cafe Bourbon, London - Cafe Bourbon, Athens - Cafe Bourbon, London - Cafe Acaia, Athens - Cafe Catuai, Sao Paulo - Cafe Catuai, Athens - Cafe Catuai.

}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataFreq)
DataFreq
}

\keyword{Data set}
