\name{DataMix}

\alias{DataMix}

\docType{data}

\title{Mixed data set.}

\description{Simulated set of mixed data on consumption of coffee.}

\usage{data(DataMix)}

\format{Data set with 10 rows and 7 columns. Being 10 observations described by 7 variables: Cooperatives/Tasters, Average grades given to analyzed coffees, Years of work as a taster, Taster with technical training, Taster exclusively dedicated, Average frequency of the coffees Classified as special, Average frequency of the coffees as commercial.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataMix)
DataMix
}

\keyword{Data set}
