\name{DataQuali}
 
\alias{DataQuali}

\docType{data}

\title{Qualitative data set}

\description{Set simulated of qualitative data on consumption of coffee.}

\usage{data(DataQuali)}

\format{Data set simulated with 12 rows and 6 columns. Being 12 observations described by 6 variables: Sex, Age, Smoker, Marital status, Sportsman, Study.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataQuali)
DataQuali
}

\keyword{Data set}
