\name{DataQuan}

\alias{DataQuan}

\docType{data}

\title{Quantitative data set}

\description{Set simulated of quantitative data on grades given to some sensory characteristics of coffees.}

\usage{data(DataQuan)}

\format{Data set with 6 rows and 11 columns. Being 6 observations described by 11 variables: Coffee, Chocolate, Caramelised, Ripe, Sweet, Delicate, Nutty, Caramelised, Chocolate, Spicy, Caramelised.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataQuan) 
DataQuan
}

\keyword{Data set}
